feat: prepare for release

+++ toktrail/adapters/base.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
+from decimal import Decimal
 from pathlib import Path
 from typing import Protocol
 
@@ -32,19 +33,19 @@ class SourceSessionSummary:
     source_paths: tuple[str, ...] = ()
 
     @property
-    def source_cost_usd(self) -> float:
+    def source_cost_usd(self) -> Decimal:
         return self.costs.source_cost_usd
 
     @property
-    def actual_cost_usd(self) -> float:
+    def actual_cost_usd(self) -> Decimal:
         return self.costs.actual_cost_usd
 
     @property
-    def virtual_cost_usd(self) -> float:
+    def virtual_cost_usd(self) -> Decimal:
         return self.costs.virtual_cost_usd
 
     @property
-    def savings_usd(self) -> float:
+    def savings_usd(self) -> Decimal:
         return self.costs.savings_usd
 
     @property
