+++ toktrail/api/models.py
@@ -174,19 +174,19 @@ class SourceSessionSummary:
     source_paths: tuple[str, ...] = ()
 
     @property
-    def source_cost_usd(self) -> float:
+    def source_cost_usd(self) -> Decimal:
         return self.costs.source_cost_usd
 
     @property
-    def actual_cost_usd(self) -> float:
+    def actual_cost_usd(self) -> Decimal:
         return self.costs.actual_cost_usd
 
     @property
-    def virtual_cost_usd(self) -> float:
+    def virtual_cost_usd(self) -> Decimal:
         return self.costs.virtual_cost_usd
 
     @property
-    def savings_usd(self) -> float:
+    def savings_usd(self) -> Decimal:
         return self.costs.savings_usd
 
     @property
